Display the bookmark in the contained components area
This involved splitting the DocumentCollectionHierarchyComponent into 2 separate components, which allows for the removal of a number of conditionals
